seperated data, events, other functionality to different classes, added click , hover functionality
